rename visitor banner to referer
