+++ js/main.js
@@ -9,7 +9,7 @@ import Modernizr from 'modernizr-mq';
 import countdown from './modules/countdown';
 import uiActions from './modules/ui-actions';
 import schedule from './modules/schedule';
-import visitor from './modules/visitor-banner';
+import referer from './modules/referer';
 import audioPlayer from './modules/audio-player';
 
 const APP = {
